feat(factions): provide a coven and dark coven faction to players
The Coven faction is what witches will join initially. This also lays the groundwork for a Capability to track their progress in joining the faction.

This also adds the Dark Coven, for players who complete the (future) Ritual of Malice or who piss off the factions.

Note that faction icons, badges, and resource bars are not currently present.  The resource bars have a bug (see todo in each Resource class), and none of these textures have been designed, yet, anyways.

Resolves #17.